slack_bot/requirements.txt
slack-bolt>=1.18.0
aiohttp>=3.8.0
nest_asyncio>=1.5.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-aws>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
openai>=1.0.0
langchain-mcp-adapters>=0.1.0
httpx>=0.27.0
pydantic>=2.0.0
mcp-snowflake-server>=0.1.0
boto3>=1.26.0 